requirements.txt
numpy==1.16.4
h5py==2.10.0
opencv_python==4.1.0.25
scipy==1.5.2
scikit_image==0.17.2
Pillow==8.1.2
tqdm==4.56.2
